_requirements.txt
amqp==5.2.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
billiard==4.2.0
blinker==1.4
celery==5.3.6
certifi==2024.7.4
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloud-init==24.3.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==4.2.14
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
filelock==3.16.1
flower==2.0.1
httplib2==0.20.2
humanize==4.10.0
hyperlink==21.0.0
idna==3.7
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpath==0.82.2
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kombu==5.3.7
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.0.1
more-itertools==8.10.0
netifaces==0.11.0
numpy==2.1.1
oauthlib==3.2.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pika==1.3.2
pipenv==2024.2.0
platformdirs==4.3.6
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycryptodome==3.20.0
pycurl==7.44.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==5.4.1
redis==5.0.1
requests==2.32.3
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
systemd-python==234
tornado==6.4.1
Twisted==22.1.0
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.2
vine==5.1.0
virtualenv==20.27.0
wadllib==1.3.6
watchfiles==0.21.0
wcwidth==0.2.13
zipp==1.0.0
zope.interface==5.4.0
